+++ cmd/serve.go
@@ -3,8 +3,10 @@ package cmd
 import (
 	"fmt"
 	"log"
+	"net/url"
 	"time"
 
+	"github.com/mono83/slf/wd"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
@@ -19,6 +21,7 @@ var serveCmd = &cobra.Command{
 	Use:   "serve",
 	Short: "Starts http handler for the skins system",
 	Run: func(cmd *cobra.Command, args []string) {
+		// TODO: this is a mess, need to organize this code somehow to make services initialization more compact
 		logger, err := bootstrap.CreateLogger(viper.GetString("statsd.addr"), viper.GetString("sentry.dsn"))
 		if err != nil {
 			log.Fatal(fmt.Printf("Cannot initialize logger: %v", err))
@@ -52,15 +55,32 @@ var serveCmd = &cobra.Command{
 			return
 		}
 
-		texturesStorage := mojangtextures.NewInMemoryTexturesStorage()
-		texturesStorage.Start()
-		mojangTexturesProvider := &mojangtextures.Provider{
-			Logger: logger,
-			UuidsProvider: &mojangtextures.BatchUuidsProvider{
+		var uuidsProvider mojangtextures.UuidsProvider
+		preferredUuidsProvider := viper.GetString("mojang_textures.uuids_provider.driver")
+		if preferredUuidsProvider == "remote" {
+			remoteUrl, err := url.Parse(viper.GetString("mojang_textures.uuids_provider.url"))
+			if err != nil {
+				logger.Emergency("Unable to parse remote url :err", wd.ErrParam(err))
+				return
+			}
+
+			uuidsProvider = &mojangtextures.RemoteApiUuidsProvider{
+				Url:    *remoteUrl,
+				Logger: logger,
+			}
+		} else {
+			uuidsProvider = &mojangtextures.BatchUuidsProvider{
 				IterationDelay: time.Duration(viper.GetInt("queue.loop_delay")) * time.Millisecond,
 				IterationSize:  viper.GetInt("queue.batch_size"),
 				Logger:         logger,
-			},
+			}
+		}
+
+		texturesStorage := mojangtextures.NewInMemoryTexturesStorage()
+		texturesStorage.Start()
+		mojangTexturesProvider := &mojangtextures.Provider{
+			Logger:        logger,
+			UuidsProvider: uuidsProvider,
 			TexturesProvider: &mojangtextures.MojangApiTexturesProvider{
 				Logger: logger,
 			},